refactoring into smaller pieces
