+++ deeplearning4j-keras/src/main/java/org/deeplearning4j/keras/NeuralNetworkReader.java
@@ -1,0 +1,27 @@
+package org.deeplearning4j.keras;
+
+import lombok.extern.slf4j.Slf4j;
+import org.deeplearning4j.nn.modelimport.keras.InvalidKerasConfigurationException;
+import org.deeplearning4j.nn.modelimport.keras.KerasModelImport;
+import org.deeplearning4j.nn.modelimport.keras.UnsupportedKerasConfigurationException;
+import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
+
+import java.io.IOException;
+
+@Slf4j
+public class NeuralNetworkReader {
+
+    public MultiLayerNetwork readNeuralNetwork(EntryPointFitParameters entryPointFitParameters)
+            throws IOException, InvalidKerasConfigurationException, UnsupportedKerasConfigurationException {
+
+        MultiLayerNetwork multiLayerNetwork;
+        if (KerasModelType.SEQUENTIAL.equals(entryPointFitParameters.getType())) {
+            multiLayerNetwork = KerasModelImport.importKerasSequentialModelAndWeights(entryPointFitParameters.getModelFilePath());
+            multiLayerNetwork.init();
+        } else {
+            throw new RuntimeException("Model type unsupported! (" + entryPointFitParameters.getType() + ")");
+        }
+
+        return multiLayerNetwork;
+    }
+}
